Fixing some problems

+++ products/models.py
@@ -45,11 +45,9 @@ class Product(models.Model):
     image_url = models.URLField(max_length=1024, null=True, blank=True)
     image = models.ImageField(null=True, blank=True)
     price = models.DecimalField(max_digits=6,decimal_places=2,null=True,blank=True,default=0.00)
-    stock_level = models.IntegerField()
+    stock_level = models.PositiveIntegerField()
     in_stock = models.BooleanField(default=True)
-    rating = models.DecimalField(
-        max_digits=6,
-        decimal_places=1,
+    rating = models.PositiveIntegerField(
         null=True,
         blank=True
     )
